Move Rollbar::Notifier to its own file
